--- examples/BuddyF16Llama/llama2-to-hf.py
@@ -1,24 +1,0 @@
-# ===- llama2-to-hf.py ---------------------------------------------------------
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-# ===---------------------------------------------------------------------------
-#
-# This is the convert of the llama2 model to huggingface format.
-#
-# ===---------------------------------------------------------------------------
-
-from transformers.models.llama import convert_llama_weights_to_hf
-
-convert_llama_weights_to_hf.main()
-
