--- examples/BuddyF16Test/.gitignore
@@ -1,4 +1,0 @@
-*.mlir
-*.llvm
-*.o
-^*.py
