--- examples/BuddyF16Test/bp16_f2b.cpp
@@ -1,3 +1,0 @@
-extern "C" {
-#include "llvm/compiler-rt/lib/builtins/truncsfbf2.c"
-}
