--- examples/BuddyF16Test/fp16_f2h.cpp
@@ -1,3 +1,0 @@
-extern "C" {
-#include "llvm/compiler-rt/lib/builtins/truncsfhf2.c"
-}
