--- examples/BuddyF16Test/fp16_h2f.cpp
@@ -1,3 +1,0 @@
-extern "C" {
-#include "llvm/compiler-rt/lib/builtins/extendhfsf2.c"
-}
