+++ examples/lit.cfg.py
@@ -42,8 +42,6 @@ config.excludes = [
     'BuddyBert',
     'BuddyMobileNetV3',
     'BuddyResNet18',
-    'BuddyF16Test',
-    'BuddyF16Llama',
     'ConvOpt',
     'DAPDialect',
     'DIPDialect',
